motor_controller: factor out common CAN functionality
This is so it can be brought out into it's own functionality.

Compiles, but not properly tested. Need to specifically check the
following:
- message reception
- services
- restarts work
